+++ backend/jobs/index.js
@@ -1,4 +1,4 @@
-const updateLeaderboards = require("./updateLeaderboards");
-const deleteOldLogs = require("./deleteOldLogs");
+const updateLeaderboards = require("./update-leaderboards");
+const deleteOldLogs = require("./delete-old-logs");
 
 module.exports = [updateLeaderboards, deleteOldLogs];
